Add WantedlyCompanyWriter
