+++ lib/wantedly_company.js
@@ -89,7 +89,28 @@ class Company {
 }
 
 
+class WantedlyCompanyWriter {
+  constructor(filename) {
+    this.destPath = path.join(__dirname, '..', 'data', filename);
+  }
+
+  async append(companies) {
+    for await (const company of companies) {
+      const row = [
+        `"${company.title}"`,
+        `"${company.status}"`,
+        `"${company.siteUrl}"`,
+        `"${company.text}"`,
+      ].join(',') + '\n';
+
+      await fs.appendFile(this.destPath, row);
+    }
+  };
+}
+
+
 
 module.exports = {
-  WantedlyCompanyList
+  WantedlyCompanyList,
+  WantedlyCompanyWriter,
 };
